requirements.txt
selenium==3.141.0
allure-pytest==2.8.36
pytest==6.2.2
